refactor: component 분리
