--- app/src/main/java/woowacourse/payments/feature/AddPaymentCardScreen.kt
@@ -1,172 +1,0 @@
-package woowacourse.payments.feature
-
-import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.automirrored.filled.ArrowBack
-import androidx.compose.material.icons.filled.Check
-import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.Icon
-import androidx.compose.material3.IconButton
-import androidx.compose.material3.OutlinedTextField
-import androidx.compose.material3.Text
-import androidx.compose.material3.TopAppBar
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.shadow
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.TextRange
-import androidx.compose.ui.text.input.PasswordVisualTransformation
-import androidx.compose.ui.text.input.TextFieldValue
-import androidx.compose.ui.text.input.VisualTransformation
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
-import woowacourse.payments.R
-import woowacourse.payments.ui.theme.Gray100
-import woowacourse.payments.ui.theme.Gray200
-
-@Preview(showBackground = true)
-@Composable
-fun AddPaymentCardScreen() {
-    Column(
-        modifier = Modifier.fillMaxSize()
-    ) {
-        NewCardTopBar(
-            onBackClick = { },
-            onSaveClick = { }
-        )
-
-        Spacer(modifier = Modifier.height(14.dp))
-
-        PaymentCard(
-            modifier = Modifier.align(Alignment.CenterHorizontally)
-        )
-
-        Spacer(modifier = Modifier.height(28.dp))
-
-        TextField(modifier = Modifier.fillMaxWidth(), R.string.label_card_number, R.string.placeholder_card_number, isCardNumber = true)
-        TextField(modifier = Modifier.fillMaxWidth(0.6f), R.string.label_expiry, R.string.placeholder_expiry, isExpiryDate = true)
-        TextField(modifier = Modifier.fillMaxWidth(), R.string.label_owner, R.string.placeholder_owner)
-        TextField(modifier = Modifier.fillMaxWidth(0.6f), R.string.label_pin, R.string.placeholder_pin, isPassword = true)
-    }
-}
-
-@Preview(showBackground = true)
-@Composable
-private fun TextField(
-    modifier: Modifier = Modifier,
-    label: Int = 0,
-    placeholder: Int = 0,
-    isPassword: Boolean = false,
-    isCardNumber: Boolean = false,
-    isExpiryDate: Boolean = false,
-) {
-    var value by remember { mutableStateOf(TextFieldValue("")) }
-
-    OutlinedTextField(
-        value = value,
-        onValueChange = { input ->
-            if (isCardNumber) {
-                val formatted = formatCardNumber(input.text)
-                value = TextFieldValue(
-                    text = formatted,
-                    selection = TextRange(formatted.length)
-                )
-            } else if (isExpiryDate) {
-                val formatted = formatExpiryDate(input.text)
-                value = TextFieldValue(
-                    text = formatted,
-                    selection = TextRange(formatted.length)
-                )
-            } else {
-                value = input
-            }
-        },
-        label = { Text(stringResource(label), color = Gray200) },
-        placeholder = { Text(stringResource(placeholder), color = Gray100) },
-        visualTransformation = if (isPassword) PasswordVisualTransformation() else VisualTransformation.None,
-        modifier = modifier.then(
-            Modifier.padding(start = 24.dp, end = 24.dp, bottom = 18.dp)
-        ),
-    )
-}
-
-@Composable
-private fun PaymentCard(
-    modifier: Modifier = Modifier,
-) {
-    Box(
-        contentAlignment = Alignment.CenterStart,
-        modifier = modifier
-            .shadow(8.dp)
-            .size(width = 208.dp, height = 124.dp)
-            .background(
-                color = Color(0xFF333333),
-                shape = RoundedCornerShape(5.dp),
-            )
-    ) {
-        Box(
-            modifier = Modifier
-                .padding(start = 14.dp, bottom = 10.dp)
-                .size(width = 40.dp, height = 26.dp)
-                .background(
-                    color = Color(0xFFCBBA64),
-                    shape = RoundedCornerShape(4.dp),
-                )
-        )
-    }
-}
-
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-private fun NewCardTopBar(
-    onBackClick: () -> Unit,
-    onSaveClick: () -> Unit,
-    modifier: Modifier = Modifier,
-) {
-    TopAppBar(
-        title = { Text(stringResource(R.string.title_add_card)) },
-        navigationIcon = {
-            IconButton(onClick = { onBackClick() }) {
-                Icon(
-                    imageVector = Icons.AutoMirrored.Filled.ArrowBack,
-                    contentDescription = stringResource(R.string.desc_back),
-                )
-            }
-        },
-        actions = {
-            IconButton(onClick = { onSaveClick() }) {
-                Icon(
-                    imageVector = Icons.Filled.Check,
-                    contentDescription = stringResource(R.string.desc_done),
-                )
-            }
-        },
-        modifier = modifier
-    )
-}
-
-private fun formatCardNumber(input: String): String {
-    val digits = input.filter { it.isDigit() }.take(16)
-    return digits.chunked(4).joinToString(" - ")
-}
-
-private fun formatExpiryDate(input: String): String {
-    val digits = input.filter { it.isDigit() }.take(4)
-    return digits.chunked(2).joinToString(" / ")
-}
